app/requirements.txt
FlaskFlask==2.0.3